refactor(issuance): extract tests to dedicated module
Move inline test module from mod.rs to separate tests.rs file for
better code organization and maintainability.
